+++ src/utils/index.js
@@ -18,6 +18,7 @@ export * from './getFileExtension';
 export * from './generateSettingsLabel';
 export * from './stringToWords';
 export * from './getDropdownOptions';
+export * from './getQueryParts';
 export * from './fetchJsonShemas';
 export * from './withProfileWrapper';
 export * from './getFieldsMatched';