+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 * Remove extra action buttons in profiles Settings screens (UIDATIMP-394 - UIDATIMP-397)
 * Augment RecordTypeSelector component with Incoming Record Type selection dropdown (UIDATIMP-386)
 * Job profile details: suppress the delete/trash can icon (UIDATIMP-390)
+* Job Profile Tree: Changes needed to support Static value submatches (UIPFIMP-11)
 
 ### Bugs fixed:
 * Profile Associator lists are empty when the user reloads the page with Profile Edit Form open (UIDATIMP-338)